스터디 문제 수정 / ch12 추가

+++ java-exercise/src/main/java/com/ohgiraffers/teamStudy/ts_0116_inheritance/ex_account/Account.java
@@ -25,7 +25,7 @@ public class Account {
    public double getBalance() {
       return bal;
    }
-   public double getAccountNumber() {
+   public int getAccountNumber() {
       return accnum;
    }
